pyproject.toml
[project]
name = "yoiblog"
version = "0.1.0"
description = "A modern static blog generator with WebUI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "Jinja2>=3.1",
    "Markdown>=3.4",
    "pymdown-extensions>=10.0",
    "python-frontmatter>=1.0",
    "PyYAML>=6.0",
    "Pygments>=2.15",
    "Flask>=3.0",
    "feedgen>=1.0",
    "watchdog>=3.0",
    "colorama>=0.4",
]

[project.scripts]
yoiblog = "yoiblog.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
yoiblog = [
    "themes/**/*",
    "scaffold_templates/**/*",
    "webui/static/**/*",
    "webui/templates/**/*",
]
